Refactor Tiles cog
This change makes the majority of the tile functionality Discord-
agnostic, so that it can be unit tested.

Discord-agnostic functionality is moved into a helper file.
